+++ auth-service/folder_test_unit/auth_service.test.js
@@ -1,5 +1,5 @@
 const request = require('supertest');
-const app = require('../index');
+const app = require('../app');
 const pool = require('../db');
 const bcrypt = require('bcrypt');
